table_process/data_structure/requirements.txt
# BigQuery and Google Cloud
google-cloud-bigquery==3.11.0
google-cloud-storage==2.10.0
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0

# Data processing
pandas==2.0.3
numpy==1.24.3
tqdm==4.65.0

# Environment management
python-dotenv==1.0.0
